aggiunta scheduler per reset smartphone noleggiati giornaliero

+++ 736752_lomuscio_cs2022/src/programmazione2/casoStudio/dispositivi/Smartphone.java
@@ -11,7 +11,7 @@ import java.lang.Cloneable;
  */
 public interface Smartphone extends Cloneable {
 
-	public int getIMEI();
+	public long getIMEI();
 
 	public String getModello();
