+++ 736752_lomuscio_cs2022/src/programmazione2/casoStudio/dispositivi/SmartphoneMedio.java
@@ -25,7 +25,7 @@ public class SmartphoneMedio extends SmartphoneBase {
 	 * @param nomeDispositivo
 	 * @param marca
 	 */
-	public SmartphoneMedio(int iMEI, String modello, int memoria, int ram, String processore, String risoluzione,
+	public SmartphoneMedio(long iMEI, String modello, int memoria, int ram, String processore, String risoluzione,
 			String nomeDispositivo, String marca) {
 		super(iMEI, modello, memoria, ram, processore, risoluzione, nomeDispositivo, marca);
 	}
